LPS-49142 Move
