+++ modules/portal/portal-spring-extender/src/META-INF/spring/parent/service-configurator-spring.xml
@@ -8,13 +8,13 @@
 	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 	xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd http://www.springframework.org/schema/osgi http://www.springframework.org/schema/osgi/spring-osgi.xsd"
 >
-	<bean id="com.liferay.portal.spring.extender.internal.service.ServiceConfigurator" class="com.liferay.portal.spring.extender.internal.service.impl.ServiceConfiguratorImpl">
+	<bean id="com.liferay.portal.spring.extender.internal.service.configurator.ServiceConfigurator" class="com.liferay.portal.spring.extender.internal.service.configurator.impl.ServiceConfiguratorImpl">
 		<property name="serviceComponentLocalService" ref="com.liferay.portal.service.ServiceComponentLocalService" />
 	</bean>
-	<bean id="com.liferay.portal.spring.extender.internal.service.listener.BaseServiceConfiguratorListener" class="com.liferay.portal.spring.extender.internal.service.listener.BaseServiceConfiguratorListener" abstract="true">
-		<property name="serviceConfigurator" ref="com.liferay.portal.spring.extender.internal.service.ServiceConfigurator" />
+	<bean id="com.liferay.portal.spring.extender.internal.service.configurator.listener.BaseServiceConfiguratorListener" class="com.liferay.portal.spring.extender.internal.service.configurator.listener.BaseServiceConfiguratorListener" abstract="true">
+		<property name="serviceConfigurator" ref="com.liferay.portal.spring.extender.internal.service.configurator.ServiceConfigurator" />
 	</bean>
-	<bean id="com.liferay.portal.spring.extender.internal.service.listener.StartedServiceConfiguratorListener" class="com.liferay.portal.spring.extender.internal.service.listener.StartedServiceConfiguratorListener" parent="com.liferay.portal.spring.extender.internal.service.listener.BaseServiceConfiguratorListener" />
-	<bean id="com.liferay.portal.spring.extender.internal.service.listener.StoppedServiceConfiguratorListener" class="com.liferay.portal.spring.extender.internal.service.listener.StoppedServiceConfiguratorListener" parent="com.liferay.portal.spring.extender.internal.service.listener.BaseServiceConfiguratorListener" />
+	<bean id="com.liferay.portal.spring.extender.internal.service.configurator.listener.StartedServiceConfiguratorListener" class="com.liferay.portal.spring.extender.internal.service.configurator.listener.StartedServiceConfiguratorListener" parent="com.liferay.portal.spring.extender.internal.service.configurator.listener.BaseServiceConfiguratorListener" />
+	<bean id="com.liferay.portal.spring.extender.internal.service.configurator.listener.StoppedServiceConfiguratorListener" class="com.liferay.portal.spring.extender.internal.service.configurator.listener.StoppedServiceConfiguratorListener" parent="com.liferay.portal.spring.extender.internal.service.configurator.listener.BaseServiceConfiguratorListener" />
 	<osgi:reference id="com.liferay.portal.service.ServiceComponentLocalService" availability="mandatory" filter="(original.bean=true)" interface="com.liferay.portal.service.ServiceComponentLocalService" />
 </beans>
